oom-groups.conf: Make sure init is the last to be OOM-killed
Killing init results in stopping the container so it must be the last
process to be considered for sacrificing on OOM. Make sure, it is by
setting its oom_score_adj to the lowest possible, i.e. -1000.

Also, set the same OOM protection for systemd init process. The thing
is, although systemd allows to configure oom_score_adj via unit files,
which makes oom-groups.conf mostly useless for systemd-based containers,
it does not set any protection for itself, because the system-wide init
process is never killed by the kernel. For containers, it isn't true
though, and init can be easily killed, e.g. if fork bomb is started
inside the container, so we need to explicitly protect it. The easiest
way to do it for now is adding a rule to oom-groups.conf.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>