Increase kernel.pid_max in userspace via /etc/sysctl.conf PSBM-55365
Since 3.10.0-493.vz7.25.1 the standard value of kernel.pid_max was restored to PID_MAX_DEFAULT which is equal to 32768.
On nodes with a lot of containers pids will rotate too frequently, moreover it can be not enough pids.
This patch will set the old value (kernel.pid_max=1048576) previously provided by the kernel from userspace.