vzauth: show proper error if .psa.shadow is absent #PVA-36225
According to vzauth's manpage the utility has to return
specific error code if /etc/psa/.psa.shadow doesn't exists.
This code explicitly specifies that the Plesk panel is
not installed inside a CT. VA Agent relies on this behaviour
when it needs to search for alive Plesk panel inside a CT.

However now vzauth returns only generic auth error.
That was fixed.